Module-7 trying to get tests running in docker
