+++ .dockerignore
@@ -1,5 +1,9 @@
 .env
+.env.test
+.env.template
+geckodriver.log
 .git
+.gitignore
 Dockerfile
 README.md
 .vagrant
@@ -10,3 +14,5 @@ Vagrantfile
 geckodriver.log
 logs.txt
 requirement.txt
+.venv
+__pycache__
